# perf(context): improve initializing `Context`