--- a/runtime_tests/deno/hono.test.ts
+++ b/runtime_tests/deno/hono.test.ts
@@ -1,7 +1,6 @@
 import { Context } from '../../src/context.ts'
 import { env, getRuntimeKey } from '../../src/helper/adapter/index.ts'
 import { Hono } from '../../src/hono.ts'
-import { HonoRequest } from '../../src/request.ts'
 import { assertEquals } from './deps.ts'
 
 // Test just only minimal patterns.