--- a/src/adapter/bun/server.test.ts
+++ b/src/adapter/bun/server.test.ts
@@ -1,5 +1,4 @@
 import { Context } from '../../context'
-import { HonoRequest } from '../../request'
 import { getBunServer } from './server'
 import type { BunServer } from './server'
 
@@ -7,13 +6,11 @@ describe('getBunServer', () => {
   it('Should success to pick Server', () => {
     const server = {} as BunServer
 
-    expect(
-      getBunServer(new Context(new HonoRequest(new Request('http://localhost/')), { env: server }))
-    ).toBe(server)
-    expect(
-      getBunServer(
-        new Context(new HonoRequest(new Request('http://localhost/')), { env: { server } })
-      )
-    ).toBe(server)
+    expect(getBunServer(new Context(new Request('http://localhost/'), { env: server }))).toBe(
+      server
+    )
+    expect(getBunServer(new Context(new Request('http://localhost/'), { env: { server } }))).toBe(
+      server
+    )
   })
 })