--- a/src/adapter/cloudflare-workers/conninfo.test.ts
+++ b/src/adapter/cloudflare-workers/conninfo.test.ts
@@ -1,5 +1,4 @@
 import { Context } from '../../context'
-import { HonoRequest } from '../../request'
 import { getConnInfo } from './conninfo'
 
 describe('getConnInfo', () => {
@@ -10,7 +9,7 @@ describe('getConnInfo', () => {
         'cf-connecting-ip': address,
       },
     })
-    const c = new Context(new HonoRequest(req))
+    const c = new Context(req)
 
     const info = getConnInfo(c)