--- a/src/compose.test.ts
+++ b/src/compose.test.ts
@@ -1,6 +1,5 @@
 import { compose } from './compose'
 import { Context } from './context'
-import { HonoRequest } from './request'
 import type { Params } from './router'
 
 type C = {
@@ -88,7 +87,7 @@ describe('compose with returning a promise, non-async function', () => {
 describe('Handler and middlewares', () => {
   const middleware: MiddlewareTuple[] = []
 
-  const req = new HonoRequest(new Request('http://localhost/'))
+  const req = new Request('http://localhost/')
   const c: Context = new Context(req)
 
   const mHandlerFoo = async (c: Context, next: Function) => {
@@ -124,7 +123,7 @@ describe('Handler and middlewares', () => {
 describe('compose with Context - 200 success', () => {
   const middleware: MiddlewareTuple[] = []
 
-  const req = new HonoRequest(new Request('http://localhost/'))
+  const req = new Request('http://localhost/')
   const c: Context = new Context(req)
   const handler = (c: Context) => {
     return c.text('Hello')
@@ -148,7 +147,7 @@ describe('compose with Context - 200 success', () => {
 describe('compose with Context - 404 not found', () => {
   const middleware: MiddlewareTuple[] = []
 
-  const req = new HonoRequest(new Request('http://localhost/'))
+  const req = new Request('http://localhost/')
   const onNotFound = (c: Context) => {
     return c.text('onNotFound', 404)
   }
@@ -185,7 +184,7 @@ describe('compose with Context - 404 not found', () => {
 describe('compose with Context - 401 not authorized', () => {
   const middleware: MiddlewareTuple[] = []
 
-  const req = new HonoRequest(new Request('http://localhost/'))
+  const req = new Request('http://localhost/')
   const c: Context = new Context(req)
   const handler = (c: Context) => {
     return c.text('Hello')
@@ -211,7 +210,7 @@ describe('compose with Context - 401 not authorized', () => {
 describe('compose with Context - next() below', () => {
   const middleware: MiddlewareTuple[] = []
 
-  const req = new HonoRequest(new Request('http://localhost/'))
+  const req = new Request('http://localhost/')
   const c: Context = new Context(req)
   const handler = (c: Context) => {
     const message = c.req.header('x-custom') || 'blank'
@@ -238,7 +237,7 @@ describe('compose with Context - next() below', () => {
 describe('compose with Context - 500 error', () => {
   const middleware: MiddlewareTuple[] = []
 
-  const req = new HonoRequest(new Request('http://localhost/'))
+  const req = new Request('http://localhost/')
   const c: Context = new Context(req)
 
   it('Error on handler', async () => {
@@ -303,7 +302,7 @@ describe('compose with Context - 500 error', () => {
   })
 })
 describe('compose with Context - not finalized', () => {
-  const req = new HonoRequest(new Request('http://localhost/'))
+  const req = new Request('http://localhost/')
   const c: Context = new Context(req)
   const onNotFound = (c: Context) => {
     return c.text('onNotFound', 404)