--- a/src/helper/streaming/sse.test.ts
+++ b/src/helper/streaming/sse.test.ts
@@ -1,9 +1,8 @@
 import { Context } from '../../context'
-import { HonoRequest } from '../../request'
 import { streamSSE } from '.'
 
 describe('SSE Streaming helper', () => {
-  const req = new HonoRequest(new Request('http://localhost/'))
+  const req = new Request('http://localhost/')
   let c: Context
   beforeEach(() => {
     c = new Context(req)