--- a/src/helper/streaming/stream.test.ts
+++ b/src/helper/streaming/stream.test.ts
@@ -1,5 +1,4 @@
 import { Context } from '../../context'
-import { HonoRequest } from '../../request'
 import { stream } from '.'
 
 describe('Basic Streaming Helper', () => {