This PR improves the performance for initialization of `Context`:

* Delay initialization for variables as long as possible.
* Pass `Request` object instead of `HonoRequest` and make it later if `c.req` is called.
* Some tests have been changed. But I think it is not a problem because `Context` is used in the `hono` package internally, though a type of `Context` is exported.
* It's not related to performance, but use `#` for some private properties.

### Result

With this benchmark code:

```ts
import { bench, group, run } from 'mitata'
import { Hono as HonoNext } from '../../dist/index.js'
import { Hono } from 'hono'

const req = new Request('http://localhost/')
const app = new Hono().get('/', (c) => c.text('Hi'))
const appNext = new HonoNext().get('/', (c) => c.text('Hi'))

bench('noop', () => {})

group(() => {
  bench('4.4.8', async () => {
    await app.fetch(req)
  })
  bench('next', async () => {
    await appNext.fetch(req)
  })
})

run()
```

The performance is improved.

Bun:

```
benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
4.4.8         906 ns/iter     (850 ns … 1'490 ns)    892 ns  1'303 ns  1'490 ns
next          821 ns/iter     (772 ns … 1'145 ns)    822 ns  1'094 ns  1'145 ns

summary
  next
   1.1x faster than 4.4.8
```

Node.js:

```
benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
4.4.8       1'099 ns/iter   (1'012 ns … 2'116 ns)  1'101 ns  1'715 ns  2'116 ns
next        1'055 ns/iter     (987 ns … 1'317 ns)  1'075 ns  1'274 ns  1'317 ns

summary
  next
   1.04x faster than 4.4.8
```

### The author should do the following, if applicable

- [ ] Add tests
- [ ] Run tests
- [ ] `bun run format:fix && bun run lint:fix` to format the code
- [ ] Add [TSDoc](https://tsdoc.org/)/[JSDoc](https://jsdoc.app/about-getting-started) to document the code