@yusukebe 2024-06-26
Hey @usualoma !

What do you think of this `perf`?