@usualoma 2024-06-26
@yusukebe 
Great! Let’s go with this.