@EdamAme-x 2024-06-26
looks faster!

```ts
------------------------------------------------- -----------------------------
4.4.8       1'890 ns/iter   (1'656 ns … 4'385 ns)  1'768 ns  4'319 ns  4'385 ns
next        1'563 ns/iter   (1'435 ns … 2'181 ns)  1'566 ns  2'167 ns  2'181 ns

summary
  next
   1.21x faster than 4.4.8
```